Only check for the @type annotation on tests
At the moment the Annotation check looks at all public functions with a
docblock, however at time you may want to document methods that are not
tests, for example a data provider. As these won't have the small /
medium annotations the check will fail.

This updates the check to only test docblocks that conatin @test within
them, which is the only place the annotation should appear